requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
flake8==7.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-telegram-bot==21.7
sniffio==1.3.1
